requirements.txt
# Core dependencies
telethon==1.37.0
python-dotenv==1.0.1
aiosqlite==0.20.0
sqlalchemy==2.0.35

# AI/LLM dependencies
google-genai==1.16.1
httpx==0.28.1
pillow==11.0.0

# Google Auth dependencies
google-auth==2.40.2
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Other dependencies from original requirements
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
